requirements.txt
pandas==2.3.3
openai==2.17.0
chromadb==1.4.1
PyPDF2==3.0.1
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-ai==1.55.0
streamlit==1.54.0
plotly==6.5.2
httpx==0.28.1
